Add inRubyBuildAgent global method
Equivalent to running code in the "ruby-build" agent. The configuration was
copied from the global "ruby-build" Kubernetes agent definition.

Difference from the global agent definition is that now it is version
controlled and also extensible. The extensibility in particular is required for
adding the docker image scanner to the pod.

This should simplify migration to ECR and using the new docker scanner in our
Ruby projects. Also, I hope it sets an example of using the shared libraries to
abstract away common pipeline code.

INF-1446