Copy deployment logic from the release project
Instead of it being a separate build, it's now part of whatever build triggers
it. This avoids the user having to jump between different builds to understand
the status of the deploy.

The code was changed slightly to fit the class model of the shared library.
E.g. steps are accessed through the `script` variable, because of reasons
described in [the Accessing Steps section of the Jenkins Extending with Shared
Libraries article][1].

INF-1607

[1]: https://jenkins.io/doc/book/pipeline/shared-libraries/#accessing-steps