Deploy using our internal release tool
Instead of nesting `podTemplates` or (`inPod` calls), we now provide the
necessary containers and volumes to add to the pod via
`deployer.wrapPodTemplate`. It is the library user's responsibility to wrap
their `podTemplate` or `inPod` arguments with the provided wrapper in whichever
pod the `deployer.deployOnCommentTrigger` is called. This is done, because pods
created by nested pod templates don't share a working directory, which means
that a new merge commit would have to be created in the new pod of the project
to be deployed, which is not guaranteed to be the exact same as the merge
commit that was created in the initial pod. There are differences when the
merge isn't a fast forward, or the master has changed between the two merges.

INF-1608