Remove no-op code
This sets the configuration as expected, but it's of no use - the merge commit
is created outside the scope of this library by the `checkout(scm)` call in the
project's Jenkinsfile.

Configured the git plugin's "Global Config user.name Value" and "Global Config
user.email Value" instead, through the Jenkins UI. This way all projects get
the correct values by default.

It would be possible to support something like
`checkout(deployer.wrapSCM(scm))`, and set the configuration values there, but
this is additional complexity for using the deploy tool.