Increase deploy timeout
Some deploys have gotten too close to the 5 minutes, while still deploying
successfully. For some graceful rollouts a longer duration might be expected.

Also considered removing the timeout and adding some notifications to the user
a la "This is taking long, please take a look." But I'm somewhat hesitant about
doing that, so this is good compromise for the time being.