Stop previous running jobs when deploying
There's an [issue][1] with the Datadog plugin that causes `is_deploy` to be
unreliable when multiple jobs are running concurrently. Whenever a job
finishes, it uses the latest configured value for `is_deploy` (or DD tags in
general). Because this is most common with `!squash` followed shortly by
a `!deploy` (or just push followed by a `!deploy`) then we can work around the
most common problem by terminating all already running jobs before configuring
the DD tags for a deploy build.

The previous build termination is based on [this SO answer][2]. Safe navigation
is used because the build might finish between the `null` check and the
following usage. Sleep is used to avoid spamming the build with stop commands.

[1]: https://github.com/DataDog/jenkins-datadog-plugin/issues/115
[2]: https://stackoverflow.com/a/44384169/1456578