Associate merge commits with sm-deployer GitHub user
I've attempted this many times, but this solution should actually work. It does
potentially (i.e. only if can't fast forward) modify the merge commit after
test have run against the commit etc, but that's okay, because the modification
only changes the author and the SHA, but we only need the SHA to remain the
same after we've tagged the docker image with the SHA.