Hold production lock until deploy fully finished
A failure in the following deploy to acceptance and merge to master means that
production has to be rolled back. If we don't hold the lock during these steps
that can still fail, somebody else might have acquired the production lock and
a rollback would have to wait. Rollbacks should be considered to have a higher
priority than regular deploys and this ensures that a rollback won't be waiting
behind a lock.

This also gets rid of a race where there are 2 simultaneous deploys: A and B.
If A is already deploying to production and B has been validated in acceptance
and its deployer clicks proceed, then B will be waiting for the production
lock. Once A finishes deploying to production, B will acquire the lock and will
check if the master branch hasn't changed. It hasn't because A hasn't pushed
its changes yet. This means that B can deploy to production, but will fail when
trying to push to master. With this commit, A will release the production lock
only after having pushed master and so when B gets the lock and checks if
master has changed, it finds it's changed and aborts the deploy.