Add Jenkinsfile for testing deployment
Loads the PR version of the pipeline library and deploys a dummy application to
all envs including prod, to test the deployment process before other projects
get to use the changes.

The build time argument is used to confirm that the correct version of code is
deployed. The runtime argument (TEMPLATE_VALUE) is used to confirm that correct
values are loaded from kube-envs.

INF-1644

[1]: https://stackoverflow.com/a/49112612/1456578