Remove version label after rollback
`kubectl rollout undo` doesn't seem to completely remove the label, so it has
to be removed manually, to completely reset the deployment to the previous
state in case of a rollback.

Patch command for removing labels is borrowed from [SO][1] and the patch
idempotency helper from a [Kubernetes issue][2]. `kubectl label` is luckily
idempotent with removals and doesn't need any special mucking.

INF-1392

[1]: https://stackoverflow.com/a/46550845/1456578
[2]: https://github.com/kubernetes/kubernetes/issues/58212#issuecomment-369725221