Add support for running smoke tests in all envs
Considered using kubectl exec, but this would require exposing test ports in
actual pods, for example. Plus finding a pod for a deployment is not trivial.
Using `kubectl run` we can sort of separate the test run from the actual pods
doing work.

In future, support for not specifying the `command` argument for `runInKube`
might be required (i.e. to use the default entrypoint/command instead), but for
our use case now, that's not required and so support isn't added.

INF-1665