Refactor sending slack messages
This changes behavior, a bit, by always adding the link to the PR. Previously
the link was only added to some messages. It's safer to include the link
always, so that whenever something goes wrong, there's always a way to quickly
find the related PR.

Also gets rid of the duplication of the channel selection from env.