Check PR reviews in addition to statuses
As explained in a comment, all reviews submitted by a user are included in
`script.pullRequest.reviews`, so we have to find the last ones for every user
and only use those. The last reviews are the ones that count towards
mergeability.

In some projects, it's currently possible to deploy without any positive
reviews. This change makes that impossible. To still support deploying without
positive reviews, we'd have to check the repository configuration and find the
min reviewer count from there.

INF-1649